+++ src/vivarium_gates_nutrition_optimization_child/components/risk.py
@@ -1,15 +1,18 @@
 import itertools
-from typing import Dict
+from typing import Callable, Dict
 
 import pandas as pd
 from vivarium.framework.engine import Builder
+from vivarium.framework.lookup import LookupTable
 from vivarium.framework.values import Pipeline
-from vivarium_public_health.risks import Risk
+from vivarium_public_health.risks import Risk, RiskEffect
 from vivarium_public_health.risks.data_transformations import (
     get_exposure_post_processor,
+    get_relative_risk_data,
     pivot_categorical,
 )
 from vivarium_public_health.risks.distributions import PolytomousDistribution
+from vivarium_public_health.utilities import EntityString, TargetString
 
 from vivarium_gates_nutrition_optimization_child.constants import data_keys, data_values
 
@@ -95,3 +98,97 @@ class ChildUnderweight(Risk):
             exposure = distribution.ppf(group_df["propensity"])
             exposures.append(exposure)
         return pd.concat(exposures).sort_index()
+
+
+class CGFRiskEffect(RiskEffect):
+    def __init__(self, target: str):
+        """
+        Parameters
+        ----------
+        target :
+            Type, name, and target rate of entity to be affected by risk factor,
+            supplied in the form "entity_type.entity_name.measure"
+            where entity_type should be singular (e.g., cause instead of causes).
+        """
+        self.risk = EntityString("risk_factor.child_growth_failure")
+        self.cgf_models = [
+            EntityString(f"risk_factor.{risk}")
+            for risk in [
+                data_keys.WASTING.name,
+                data_keys.UNDERWEIGHT.name,
+                data_keys.STUNTING.name,
+            ]
+        ]
+        self.target = TargetString(target)
+        self.configuration_defaults = self._get_configuration_defaults()
+
+        self.target_pipeline_name = f"{self.target.name}.{self.target.measure}"
+        self.target_paf_pipeline_name = f"{self.target_pipeline_name}.paf"
+
+    def _get_configuration_defaults(self) -> Dict[str, Dict]:
+        def get_config_by_risk(risk):
+            return {
+                f"effect_of_{risk.name}_on_{self.target.name}": {
+                    self.target.measure: RiskEffect.configuration_defaults[
+                        "effect_of_risk_on_target"
+                    ]["measure"]
+                }
+            }
+
+        config = get_config_by_risk(self.risk)
+        for sub_risk in self.cgf_models:
+            config.update(get_config_by_risk(sub_risk))
+        return config
+
+    def _get_distribution_type(self, builder: Builder) -> str:
+        return "ordered_polytomous"
+
+    def _get_risk_exposure(self, builder: Builder) -> Callable[[pd.Index], pd.Series]:
+        return {
+            risk: builder.value.get_value(f"{risk.name}.exposure") for risk in self.cgf_models
+        }
+
+    def _get_relative_risk_source(self, builder: Builder) -> LookupTable:
+        return {
+            risk: builder.lookup.build_table(
+                get_relative_risk_data(builder, risk, self.target),
+                key_columns=["sex"],
+                parameter_columns=["age", "year"],
+            )
+            for risk in self.cgf_models
+        }
+
+    def _get_population_attributable_fraction_source(self, builder: Builder) -> LookupTable:
+        paf_data = builder.data.load(f"{self.risk}.population_attributable_fraction")
+        correct_target = (paf_data["affected_entity"] == self.target.name) & (
+            paf_data["affected_measure"] == self.target.measure
+        )
+        paf_data = paf_data[correct_target].drop(
+            columns=["affected_entity", "affected_measure"]
+        )
+        return builder.lookup.build_table(
+            paf_data, key_columns=["sex"], parameter_columns=["age", "year"]
+        )
+
+    def _get_target_modifier(
+        self, builder: Builder
+    ) -> Callable[[pd.Index, pd.Series], pd.Series]:
+        def adjust_target(index: pd.Index, target: pd.Series) -> pd.Series:
+            rrs = self.relative_risk
+            exposures = self.exposure
+            for risk in self.cgf_models:
+                index_columns = ["index", risk.name]
+                rr = rrs[risk](index)
+                exposure = exposures[risk](index).reset_index()
+                exposure.columns = index_columns
+                exposure = exposure.set_index(index_columns)
+
+                relative_risk = rr.stack().reset_index()
+                relative_risk.columns = index_columns + ["value"]
+                relative_risk = relative_risk.set_index(index_columns)
+
+                effect = relative_risk.loc[exposure.index, "value"].droplevel(risk.name)
+                target *= effect
+            return target
+
+        return adjust_target