+++ src/vivarium_gates_nutrition_optimization_child/constants/paths.py
@@ -17,6 +17,7 @@ MMS_GA_SHIFT_2_DATA_DIR = RAW_DATA_DIR / "mms_gestational_age_shifts/shift2"
 WASTING_TRANSITIONS_DATA_DIR = RAW_DATA_DIR / "wasting_transition_rates"
 WASTING_TREATMENT_PARAMETERS_DIR = RAW_DATA_DIR / "wasting_treatment_parameters"
 UNDERWEIGHT_CONDITIONAL_DISTRIBUTIONS = RAW_DATA_DIR / "lookup.csv"
+CGF_PAFS = RAW_DATA_DIR / "cgf_pafs.csv"
 
 TEMPORARY_PAF_DIR = Path(
     "/mnt/team/simulation_science/costeffectiveness/auxiliary_data/GBD_2019/01_original_data/population_attributable_fraction/risk_factor/lbwsg/data/"