+++ src/vivarium_gates_nutrition_optimization_child/data/loader.py
@@ -73,12 +73,12 @@ def get_data(lookup_key: str, location: str) -> pd.DataFrame:
         data_keys.DIARRHEA.DURATION: load_duration,
         data_keys.DIARRHEA.PREVALENCE: load_prevalence_from_incidence_and_duration,
         data_keys.DIARRHEA.INCIDENCE_RATE: load_standard_gbd_2019_data_as_gbd_2021_data,
-        data_keys.DIARRHEA.REMISSION_RATE: load_remission_rate_from_duration,
+        data_keys.DIARRHEA.REMISSION_RATE: load_neonatal_deleted_remission_from_duration,
         data_keys.DIARRHEA.DISABILITY_WEIGHT: load_standard_gbd_2019_data_as_gbd_2021_data,
         data_keys.DIARRHEA.EMR: load_emr_from_csmr_and_prevalence,
-        data_keys.DIARRHEA.CSMR: load_diarrhea_csmr,
+        data_keys.DIARRHEA.CSMR: load_neonatal_deleted_csmr,
         data_keys.DIARRHEA.RESTRICTIONS: load_metadata,
-        data_keys.DIARRHEA.BIRTH_PREVALENCE: load_diarrhea_birth_prevalence,
+        data_keys.DIARRHEA.BIRTH_PREVALENCE: load_post_neonatal_birth_prevalence,
         data_keys.MEASLES.PREVALENCE: load_standard_gbd_2019_data_as_gbd_2021_data,
         data_keys.MEASLES.INCIDENCE_RATE: load_standard_gbd_2019_data_as_gbd_2021_data,
         data_keys.MEASLES.DISABILITY_WEIGHT: load_standard_gbd_2019_data_as_gbd_2021_data,
@@ -88,19 +88,20 @@ def get_data(lookup_key: str, location: str) -> pd.DataFrame:
         data_keys.LRI.DURATION: load_duration,
         data_keys.LRI.PREVALENCE: load_prevalence_from_incidence_and_duration,
         data_keys.LRI.INCIDENCE_RATE: load_standard_gbd_2019_data_as_gbd_2021_data,
-        data_keys.LRI.REMISSION_RATE: load_remission_rate_from_duration,
+        data_keys.LRI.REMISSION_RATE: load_neonatal_deleted_remission_from_duration,
         data_keys.LRI.DISABILITY_WEIGHT: load_standard_gbd_2019_data_as_gbd_2021_data,
         data_keys.LRI.EMR: load_emr_from_csmr_and_prevalence,
-        data_keys.LRI.CSMR: load_lri_csmr,
+        data_keys.LRI.CSMR: load_neonatal_deleted_csmr,
         data_keys.LRI.RESTRICTIONS: load_metadata,
         data_keys.MALARIA.DURATION: load_duration,
         data_keys.MALARIA.PREVALENCE: load_prevalence_from_incidence_and_duration,
         data_keys.MALARIA.INCIDENCE_RATE: load_standard_gbd_2019_data_as_gbd_2021_data,
-        data_keys.MALARIA.REMISSION_RATE: load_malaria_remission_rate_from_duration,
+        data_keys.MALARIA.REMISSION_RATE: load_neonatal_deleted_malaria_remission_from_duration,
         data_keys.MALARIA.DISABILITY_WEIGHT: load_standard_gbd_2019_data_as_gbd_2021_data,
         data_keys.MALARIA.EMR: load_emr_from_csmr_and_prevalence,
-        data_keys.MALARIA.CSMR: load_standard_gbd_2019_data_as_gbd_2021_data,
+        data_keys.MALARIA.CSMR: load_neonatal_deleted_csmr,
         data_keys.MALARIA.RESTRICTIONS: load_metadata,
+        data_keys.MALARIA.BIRTH_PREVALENCE: load_post_neonatal_birth_prevalence,
         data_keys.WASTING.DISTRIBUTION: load_metadata,
         data_keys.WASTING.ALT_DISTRIBUTION: load_metadata,
         data_keys.WASTING.CATEGORIES: load_metadata,
@@ -114,8 +115,11 @@ def get_data(lookup_key: str, location: str) -> pd.DataFrame:
         data_keys.STUNTING.EXPOSURE: load_gbd_2021_exposure,
         data_keys.STUNTING.RELATIVE_RISK: load_gbd_2021_rr,
         data_keys.STUNTING.PAF: load_categorical_paf,
+        data_keys.UNDERWEIGHT.DISTRIBUTION: load_metadata,
         data_keys.UNDERWEIGHT.EXPOSURE: load_underweight_exposure,
         data_keys.UNDERWEIGHT.CATEGORIES: load_metadata,
+        data_keys.UNDERWEIGHT.RELATIVE_RISK: load_gbd_2021_rr,
+        data_keys.CHILD_GROWTH_FAILURE.PAF: load_cgf_paf,
         data_keys.PEM.EMR: load_pem_emr,
         data_keys.PEM.CSMR: load_pem_csmr,
         data_keys.PEM.RESTRICTIONS: load_pem_restrictions,
@@ -241,8 +245,23 @@ def load_standard_data(key: str, location: str) -> pd.DataFrame:
 
 
 def load_standard_gbd_2019_data_as_gbd_2021_data(key: str, location: str) -> pd.DataFrame:
+    """Read in GBD 2019 data with GBD 2021 ages and years. Zero out neonatal age group data if necessary."""
     gbd_2019_data = load_standard_data(key, location)
-    return utilities.reshape_gbd_2019_data_as_gbd_2021_data(gbd_2019_data)
+    data = utilities.reshape_gbd_2019_data_as_gbd_2021_data(gbd_2019_data)
+
+    neonatal_deleted_keys = [
+        data_keys.DIARRHEA.INCIDENCE_RATE,
+        data_keys.DIARRHEA.DISABILITY_WEIGHT,
+        data_keys.LRI.INCIDENCE_RATE,
+        data_keys.LRI.DISABILITY_WEIGHT,
+        data_keys.MALARIA.INCIDENCE_RATE,
+        data_keys.MALARIA.DISABILITY_WEIGHT,
+    ]
+
+    if key in neonatal_deleted_keys:
+        data.loc[data.reset_index()["age_start"] < metadata.NEONATAL_END_AGE, :] = 0
+
+    return data
 
 
 def load_metadata(key: str, location: str):
@@ -395,7 +414,8 @@ def load_prevalence_from_incidence_and_duration(key: str, location: str) -> pd.D
     return prevalence
 
 
-def load_remission_rate_from_duration(key: str, location: str) -> pd.DataFrame:
+def load_neonatal_deleted_remission_from_duration(key: str, location: str) -> pd.DataFrame:
+    """Calculate remission rate from duration and zero out neonatal age group data."""
     try:
         cause = {
             data_keys.DIARRHEA.REMISSION_RATE: data_keys.DIARRHEA,
@@ -407,23 +427,28 @@ def load_remission_rate_from_duration(key: str, location: str) -> pd.DataFrame:
     duration = get_data(cause.DURATION, location)
     remission_rate = (-1 / step_size) * np.log(1 - step_size / duration)
 
-    if key == data_keys.DIARRHEA.REMISSION_RATE:
-        remission_rate.loc[
-            remission_rate.index.get_level_values("age_start") < metadata.NEONATAL_END_AGE, :
-        ] = 0
+    remission_rate.loc[
+        remission_rate.index.get_level_values("age_start") < metadata.NEONATAL_END_AGE, :
+    ] = 0
     return remission_rate
 
 
-def load_malaria_remission_rate_from_duration(key: str, location: str) -> pd.DataFrame:
-    """Return 1 / duration."""
+def load_neonatal_deleted_malaria_remission_from_duration(
+    key: str, location: str
+) -> pd.DataFrame:
+    """Return 1 / duration with zero'd out neonatal age groups."""
     try:
         cause = {
             data_keys.MALARIA.REMISSION_RATE: data_keys.MALARIA,
         }[key]
     except KeyError:
         raise ValueError(f"Unrecognized key {key}")
+
     duration = get_data(cause.DURATION, location)
-    return 1 / duration
+    data = 1 / duration
+    data.loc[data.reset_index()["age_start"] < metadata.NEONATAL_END_AGE, :] = 0
+
+    return data
 
 
 def load_emr_from_csmr_and_prevalence(key: str, location: str) -> pd.DataFrame:
@@ -443,6 +468,37 @@ def load_emr_from_csmr_and_prevalence(key: str, location: str) -> pd.DataFrame:
 
     if key == data_keys.DIARRHEA.EMR:
         data.loc[data.index.get_level_values("age_start") < metadata.NEONATAL_END_AGE, :] = 0
+    return data
+
+
+def load_neonatal_deleted_csmr(key: str, location: str) -> pd.DataFrame:
+    """Get GBD 2019 CSMR data with 2021 age groups and zero out neonatal age groups."""
+    allowed_keys = [data_keys.DIARRHEA.CSMR, data_keys.LRI.CSMR, data_keys.MALARIA.CSMR]
+    if key not in allowed_keys:
+        raise ValueError(f"Unrecognized key {key}")
+
+    data = load_standard_gbd_2019_data_as_gbd_2021_data(key, location)
+    data.loc[data.reset_index()["age_start"] < metadata.NEONATAL_END_AGE, :] = 0
+    return data
+
+
+def load_post_neonatal_birth_prevalence(key: str, location: str) -> pd.DataFrame:
+    """Return post neonatal data (1 month to 6 months) as birth prevalence."""
+    try:
+        cause = {
+            data_keys.DIARRHEA.BIRTH_PREVALENCE: data_keys.DIARRHEA,
+            data_keys.MALARIA.BIRTH_PREVALENCE: data_keys.MALARIA,
+        }[key]
+    except KeyError:
+        raise ValueError(f"Unrecognized key {key}")
+
+    prevalence = get_data(cause.PREVALENCE, location)
+    is_post_neonatal = np.isclose(
+        prevalence.reset_index()["age_start"], metadata.NEONATAL_END_AGE
+    )
+    post_neonatal_prevalence = prevalence[is_post_neonatal]
+    data = post_neonatal_prevalence.droplevel(["age_start", "age_end"])
+
     return data
 
 
@@ -519,7 +575,7 @@ def load_gbd_2021_rr(key: str, location: str) -> pd.DataFrame:
     entity_key = EntityKey(key)
     entity = utilities.get_gbd_2021_entity(entity_key)
 
-    data = utilities.get_data(
+    raw_data = utilities.get_data(
         entity_key,
         entity,
         location,
@@ -528,8 +584,8 @@ def load_gbd_2021_rr(key: str, location: str) -> pd.DataFrame:
         metadata.AGE_GROUP.GBD_2021,
         metadata.GBD_2021_ROUND_ID,
     )
-    data = utilities.process_relative_risk(
-        data,
+    raw_data = utilities.process_gbd_2021_relative_risk(
+        raw_data,
         entity_key,
         entity,
         location,
@@ -537,16 +593,47 @@ def load_gbd_2021_rr(key: str, location: str) -> pd.DataFrame:
         metadata.AGE_GROUP.GBD_2021,
     )
 
+    inc = raw_data.query('affected_measure == "incidence_rate"')
+    csmr = raw_data.query('affected_measure == "cause_specific_mortality_rate"')
+    emr = csmr.droplevel("affected_measure") / inc.droplevel("affected_measure")
+    emr["affected_measure"] = "excess_mortality_rate"
+    emr = emr.set_index("affected_measure", append=True).reorder_levels(inc.index.names)
+
+    data = pd.concat([inc, emr])
+
     if key == data_keys.STUNTING.RELATIVE_RISK:
         # Remove neonatal relative risks
         neonatal_age_ends = data.index.get_level_values("age_end").unique()[:2]
         data.loc[data.index.get_level_values("age_end").isin(neonatal_age_ends)] = 1.0
-    elif key == data_keys.WASTING.RELATIVE_RISK:
-        # Remove relative risks for simulants under 6 months
-        data.loc[
-            data.index.get_level_values("age_end") <= data_values.WASTING.DYNAMIC_START_AGE
-        ] = 1.0
     return data
+
+
+def load_cgf_paf(key: str, location: str) -> pd.DataFrame:
+    data = pd.read_csv(paths.CGF_PAFS, index_col=0)
+
+    # add age start and age end data instead of age group name
+    age_bins = get_data(data_keys.POPULATION.AGE_BINS, location).reset_index()
+    age_bins["age_group_name"] = age_bins["age_group_name"].str.lower().str.replace(" ", "_")
+    age_start_map = dict(zip(age_bins["age_group_name"], age_bins["age_start"]))
+    age_end_map = dict(zip(age_bins["age_group_name"], age_bins["age_end"]))
+    data["age_start"] = data["age_group_name"].map(age_start_map)
+    data["age_end"] = data["age_group_name"].map(age_end_map)
+    data = data.drop(["age_group_name", "location_id"], axis=1)
+
+    # duplicate data for 1990 to 2022
+    year_list = list(range(1990, 2023))
+    years = pd.DataFrame({"year_start": year_list}).set_index(pd.Index([1] * len(year_list)))
+    data = data.set_index(pd.Index([1] * len(data))).join(years)
+    data["year_end"] = data["year_start"] + 1
+
+    # Capitalize Sex
+    data["sex"] = data["sex"].str.capitalize()
+
+    # define index
+    data = data.set_index(
+        metadata.ARTIFACT_INDEX_COLUMNS + ["affected_entity", "affected_measure"]
+    )
+    return data.sort_index()
 
 
 def load_pem_disability_weight(key: str, location: str) -> pd.DataFrame:
@@ -884,6 +971,24 @@ def load_sids_csmr(key: str, location: str) -> pd.DataFrame:
         raise ValueError(f"Unrecognized key {key}")
 
 
+def load_neonatal_lri_csmr(key: str, location: str) -> pd.DataFrame:
+    if key != data_keys.AFFECTED_UNMODELED_CAUSES.NEONATAL_LRI_CSMR:
+        raise ValueError(f"Unrecognized key {key}")
+
+    data = load_standard_gbd_2019_data_as_gbd_2021_data(data_keys.LRI.CSMR, location)
+    data.loc[data.index.get_level_values("age_start") >= metadata.NEONATAL_END_AGE, :] = 0
+    return data
+
+
+def load_neonatal_diarrhea_csmr(key: str, location: str) -> pd.DataFrame:
+    if key != data_keys.AFFECTED_UNMODELED_CAUSES.NEONATAL_DIARRHEAL_DISEASES_CSMR:
+        raise ValueError(f"Unrecognized key {key}")
+
+    data = load_standard_gbd_2019_data_as_gbd_2021_data(data_keys.DIARRHEA.CSMR, location)
+    data.loc[data.index.get_level_values("age_start") >= metadata.NEONATAL_END_AGE, :] = 0
+    return data
+
+
 def load_intervention_distribution(key: str, location: str) -> str:
     try:
         return {
@@ -1190,57 +1295,6 @@ def load_maternal_bmi_anemia_excess_shift(key: str, location: str) -> pd.DataFra
         ["affected_entity", "affected_measure", "parameter"], append=True
     ).sort_index()
     return excess_shift
-
-
-def load_neonatal_lri_csmr(key: str, location: str) -> pd.DataFrame:
-    if key != data_keys.AFFECTED_UNMODELED_CAUSES.NEONATAL_LRI_CSMR:
-        raise ValueError(f"Unrecognized key {key}")
-
-    data = load_standard_gbd_2019_data_as_gbd_2021_data(data_keys.LRI.CSMR, location)
-    data.loc[data.index.get_level_values("age_start") >= metadata.NEONATAL_END_AGE, :] = 0
-    return data
-
-
-def load_lri_csmr(key: str, location: str) -> pd.DataFrame:
-    if key != data_keys.LRI.CSMR:
-        raise ValueError(f"Unrecognized key {key}")
-
-    data = load_standard_gbd_2019_data_as_gbd_2021_data(data_keys.LRI.CSMR, location)
-    data.loc[data.index.get_level_values("age_start") < metadata.NEONATAL_END_AGE, :] = 0
-    return data
-
-
-def load_diarrhea_csmr(key: str, location: str) -> pd.DataFrame:
-    if key != data_keys.DIARRHEA.CSMR:
-        raise ValueError(f"Unrecognized key {key}")
-
-    data = load_standard_gbd_2019_data_as_gbd_2021_data(data_keys.DIARRHEA.CSMR, location)
-    data.loc[data.index.get_level_values("age_start") < metadata.NEONATAL_END_AGE, :] = 0
-    return data
-
-
-def load_neonatal_diarrhea_csmr(key: str, location: str) -> pd.DataFrame:
-    if key != data_keys.AFFECTED_UNMODELED_CAUSES.NEONATAL_DIARRHEAL_DISEASES_CSMR:
-        raise ValueError(f"Unrecognized key {key}")
-
-    data = load_standard_gbd_2019_data_as_gbd_2021_data(data_keys.DIARRHEA.CSMR, location)
-    data.loc[data.index.get_level_values("age_start") >= metadata.NEONATAL_END_AGE, :] = 0
-    return data
-
-
-def load_diarrhea_birth_prevalence(key: str, location: str) -> pd.DataFrame:
-    if key != data_keys.DIARRHEA.BIRTH_PREVALENCE:
-        raise ValueError(f"Unrecognized key {key}")
-
-    prevalence = get_data(data_keys.DIARRHEA.PREVALENCE, location)
-    post_neonatal = prevalence.loc[
-        (prevalence.index.get_level_values("age_start") >= metadata.NEONATAL_END_AGE)
-        ### I think this is just supposed to pull late neonatal, and the bin sizes have changed.
-        & (prevalence.index.get_level_values("age_start") < 0.5)
-    ]
-    data = post_neonatal.droplevel(["age_start", "age_end"])
-
-    return data
 
 
 def reshape_to_vivarium_format(df, location):